+++ tests/Unit/DSC_NetAdapterState.Tests.ps1
@@ -38,11 +38,13 @@ try
         $netAdapterEnabled = [PSCustomObject]@{
             Name        = 'Ethernet'
             AdminStatus = [Microsoft.PowerShell.Cmdletization.GeneratedTypes.NetAdapter.NET_IF_ADMIN_STATUS]::Up
+            State       = 'Enabled'
         }
 
         $netAdapterDisabled = [PSCustomObject]@{
             Name        = 'Ethernet'
             AdminStatus = [Microsoft.PowerShell.Cmdletization.GeneratedTypes.NetAdapter.NET_IF_ADMIN_STATUS]::Down
+            State       = 'Disabled'
         }
 
         $netAdapterUnsupported = [PSCustomObject]@{
@@ -132,6 +134,7 @@ try
                     $netAdapterEnabled
                 }
                 Mock -CommandName Disable-NetAdapter
+                Mock -CommandName Enable-NetAdapter
 
                 It 'Should not throw an exception' {
                     { Set-TargetResource @setTargetResourceEnabled } | Should -Not -Throw
@@ -140,6 +143,7 @@ try
                 It 'Should not call Disable-NetAdapter' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                     Assert-MockCalled -CommandName Disable-NetAdapter -Exactly -Times 0
+                    Assert-MockCalled -CommandName Enable-NetAdapter -Exactly -Times 0
                 }
             }
 
@@ -165,8 +169,8 @@ try
                 Mock -CommandName Get-NetAdapter -MockWith {
                     $netAdapterDisabled
                 }
-                Mock -CommandName Enable-NetAdapter
                 Mock -CommandName Disable-NetAdapter
+                Mock -CommandName Enable-NetAdapter
 
                 It 'Should not throw an exception' {
                     { Set-TargetResource @setTargetResourceDisabled } | Should -Not -Throw
@@ -174,6 +178,7 @@ try
 
                 It 'Should not call Enable-NetAdapter' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
+                    Assert-MockCalled -CommandName Disable-NetAdapter -Exactly -Times 0
                     Assert-MockCalled -CommandName Enable-NetAdapter -Exactly -Times 0
                 }
             }