+++ CHANGELOG.md
@@ -32,6 +32,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     support for changing proxy settings for current user.
     BREAKING CHANGE: Added support for configuring proxy settings for a user
     account by adding `Target` parameter - Fixes [Issue #423](https://github.com/dsccommunity/NetworkingDsc/issues/423).
+- Updated .github issue templates to standard - Fixes [Issue #508](https://github.com/dsccommunity/NetworkingDsc/issues/508).
+- Added Create_ChangeLog_GitHub_PR task to publish stage of build pipeline.
+- Added SECURITY.md.
+- Updated pipeline Deploy_Module anb Code_Coverage jobs to use ubuntu-latest
+  images - Fixes [Issue #508](https://github.com/dsccommunity/NetworkingDsc/issues/508).
+- Updated pipeline unit tests and integration tests to use Windows Server 2019 and
+  Windows Server 2022 images - Fixes [Issue #507](https://github.com/dsccommunity/NetworkingDsc/issues/507).
+- NetAdapterState
+  - Added a new message when setting the state of an adapter.
+
+### Fixed
+
+- Fixed pipeline by replacing the GitVersion task in the `azure-pipelines.yml`
+  with a script.
+- NetAdapterState
+  - Fixed so that the resource is idempotent so that `Enable-NetAdapter` and
+    `Disable-NetAdapter` are only called when change is required.
+- NetAdapterLso
+  - Fixed integration tests so that they will be skipped if a network adapter
+    with NDIS version 6 or greater is not available.
 
 ## [8.2.0] - 2020-10-16